Fix Windows/MSVC C4244 narrowing warnings: float/double to int
Add static_cast<> around float/double values narrowed to an
integer-family type (C4244). Left size_t-alias-to-int narrowing
(e.g. Dali::VectorBase::SizeType <-> uint32_t) untouched, since
that is an int<->int conversion needing separate case-by-case review.

+++ dali/internal/event/actors/camera-actor-impl.cpp
@@ -614,7 +614,7 @@ bool CameraActor::BuildPickingRay(const Vector2& screenCoordinates,
                                   Vector4&       rayDirection,
                                   bool           isDrawnOnFramebuffer)
 {
-  Viewport viewport(0, 0, screenSize.x, screenSize.y);
+  Viewport viewport(0, 0, static_cast<int32_t>(screenSize.x), static_cast<int32_t>(screenSize.y));
   return BuildPickingRay(screenCoordinates, viewport, rayOrigin, rayDirection, isDrawnOnFramebuffer);
 }
